src/requirements/prod.txt
-r common.txt
django-launch-library==2.6.13